Added Authorization req.

+++ routers/ad.router.ts
@@ -1,6 +1,9 @@
 import {Router} from "express";
 import {AdRecord} from "../records/ad.record";
-import {HttpError} from "../utils/errors";
+import {HttpError, ValidationError} from "../utils/errors";
+import checkAuth from "../utils/check-auth"
+import {AuthReq} from "../types";
+
 
 export const adRouter = Router();
 
@@ -37,21 +40,28 @@ adRouter.get('/', async (req, res) => {
             ads,
         });
     })
-    .post('/', async (req, res) => {
+    .use(checkAuth)
+    .post('/', async (req:AuthReq, res) => {
         const newAd = new AdRecord(req.body);
         await newAd.insert();
+        if  (newAd.creatorId !== req.userData.user.userId){
+            throw new ValidationError('You are not allowed to Added this ad.')
+        }
 
         res.status(201)
             .json({
                 newAd,
             });
     })
-    .patch('/:aId', async (req, res) => {
+    .patch('/:aId', async (req:AuthReq, res) => {
         const {aId} = req.params;
         const body = req.body;
         const ad = await AdRecord.getOne(aId);
         if (ad === null) {
             throw new HttpError('Could not find a ad for the provided id.', 404);
+        }
+        if  (ad.creatorId !== req.userData.user.userId){
+            throw new ValidationError('You are not allowed to edit this ad.')
         }
         if (ad.id !== body.id) {
             throw new HttpError('id must be provided.', 400);
@@ -63,11 +73,15 @@ adRouter.get('/', async (req, res) => {
         });
 
     })
-    .delete('/:id', async (req, res) => {
+    .delete('/:id', async (req:AuthReq, res) => {
         const ad = await AdRecord.getOne(req.params.id);
         if (!ad) {
-            throw new HttpError('No such ad.', 404);
+            throw new HttpError('No found ad.', 404);
         }
+        if  (ad.creatorId !== req.userData.user.userId){
+            throw new ValidationError('You are not allowed to delete this ad.')
+        }
+
 
         await ad.delete();
         res.status(200).json({message: "Delete ad successfully"});
